Update landVehicle.java
Removing reference to  // super(seats, headlights);
In tnhis context, this means that we want to call the constructor for the parent clsas (the base class) vehicle, that accepts two parameters, the number of seats and the number of headlights,.
There is no constructor in class vehicle. In fact there are no constructors at all in class vehicle, so calling  super(seats, headlights); will cause a compiler error as no such consrtuctor exists.